Refactor: extracted favourites stuff from property controller
