+++ src/main/java/cmpt276/pg6/realtorest/controllers/AdminController.java
@@ -85,7 +85,6 @@ public class AdminController {
         List<Admin> adminList = adminRepo.findByEmailAndPassword(email, password);
         if (adminList.isEmpty()) {
             // If no user that matches the email and password is found, return to the login page
-            // TODO Add a message to the login page that says "Invalid email or password"
             model.addAttribute("errorMessage", "Invalid Credentials Entered.");
             return "users/adminlogin";
         } else {
