+++ src/main/java/cmpt276/pg6/realtorest/controllers/PropertyController.java
@@ -2,14 +2,10 @@ package cmpt276.pg6.realtorest.controllers;
 
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -18,7 +14,6 @@ import org.springframework.web.bind.annotation.RequestParam;
 import cmpt276.pg6.realtorest.models.Property;
 import cmpt276.pg6.realtorest.models.PropertyRepository;
 import cmpt276.pg6.realtorest.models.User;
-import cmpt276.pg6.realtorest.models.UserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
@@ -28,15 +23,8 @@ public class PropertyController {
     @Autowired
     private PropertyRepository propertyRepo;
 
-    @Autowired
-    private UserRepository userRepo;
-
     public void setPropertyRepo(PropertyRepository propertyRepo) {
         this.propertyRepo = propertyRepo;
-    }
-
-    public void setUserRepo(UserRepository userRepo) {
-        this.userRepo = userRepo;
     }
 
     /**
@@ -105,24 +93,6 @@ public class PropertyController {
         model.addAttribute("properties", properties);
         System.out.println("Fetched properties: " + properties.size());
         return "propertyListing";
-    }
-
-    @GetMapping("/favourites")
-    public String showFavourites(HttpServletRequest request, HttpSession session, Model model) {
-        User sessionUser = (User) session.getAttribute("session_user");
-        Integer userId = sessionUser != null ? sessionUser.getUid() : null;
-
-        if (userId != null) {
-            Optional<User> userOptional = userRepo.findById(userId);
-            if (userOptional.isPresent()) {
-                User user = userOptional.get();
-                Set<Property> favouriteProperties = user.getFavouriteProperties();
-                model.addAttribute("favouriteProperties", favouriteProperties);
-                model.addAttribute("user", user); // Add this line
-                return "favourites";
-            }
-        }
-        return "login";
     }
 
     // Show edit property page
@@ -203,44 +173,4 @@ public class PropertyController {
         propertyRepo.deleteAll();
         return "redirect:" + redirectUrl;
     }
-
-    //Add property to favourites
-    @PostMapping("/add-favourite/{propertyId}")
-    public ResponseEntity<String> addToFavourites(@PathVariable Integer propertyId, HttpServletRequest request,
-        HttpSession session) {
-        User sessionUser = (User) session.getAttribute("session_user");
-        Integer userId = sessionUser != null ? sessionUser.getUid() : null;
-
-        if (userId != null) {
-            User user = userRepo.findById(userId).orElse(null);
-            Property property = propertyRepo.findById(propertyId).orElse(null);
-
-            if (user != null && property != null) {
-                user.getFavouriteProperties().add(property);
-                userRepo.save(user);
-                return ResponseEntity.ok("Property added to favourites successfully");
-            }
-        }
-        return ResponseEntity.badRequest().body("User or Property not found");
-    }
-
-    //Remove property from favourites
-    @DeleteMapping("/remove-favourite/{propertyId}")
-    public ResponseEntity<String> removeFromFavourites(@PathVariable Integer propertyId, HttpServletRequest request,
-        HttpSession session) {
-        User sessionUser = (User) session.getAttribute("session_user");
-        Integer userId = sessionUser != null ? sessionUser.getUid() : null;
-
-        if (userId != null) {
-            Optional<User> userOptional = userRepo.findById(userId);
-            if (userOptional.isPresent()) {
-                User user = userOptional.get();
-                user.getFavouriteProperties().removeIf(property -> property.getPid() == propertyId);
-                userRepo.save(user);
-                return ResponseEntity.ok("Property removed from favourites successfully");
-            }
-        }
-        return ResponseEntity.badRequest().body("User not found");
-    }
-
 }
